% Auto-generated: do not edit by hand
\name{''DashAlert}

\alias{''DashAlert}

\title{DashAlert component}

\description{

}

\usage{
''DashAlert(id=NULL, buttons=NULL, duration=NULL, is_open=NULL,
message=NULL, output_target=NULL, title=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{buttons}{List of named lists. }

\item{duration}{Numeric. }

\item{is_open}{Logical. }

\item{message}{A list of or a singular dash component, string or number. }

\item{output_target}{Logical | numeric | character | named list | unnamed list. }

\item{title}{A list of or a singular dash component, string or number. }

\item{value}{Character | numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

